# apps/web/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AmountBandRow": {
        "additionalProperties": false,
        "properties": {
          "band_id": {
            "title": "Band Id",
            "type": "string"
          },
          "conversion_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversion Rate"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "label_fa": {
            "title": "Label Fa",
            "type": "string"
          },
          "max_exclusive_irr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Exclusive Irr"
          },
          "min_irr": {
            "title": "Min Irr",
            "type": "integer"
          },
          "sessions": {
            "title": "Sessions",
            "type": "integer"
          },
          "verified_revenue": {
            "title": "Verified Revenue",
            "type": "integer"
          },
          "verified_sessions": {
            "title": "Verified Sessions",
            "type": "integer"
          }
        },
        "required": [
          "band_id",
          "label_fa",
          "min_irr",
          "max_exclusive_irr",
          "sessions",
          "verified_sessions",
          "conversion_rate",
          "verified_revenue",
          "evidence_id"
        ],
        "title": "AmountBandRow",
        "type": "object"
      },
...
      "DashboardInsight": {
        "additionalProperties": false,
        "properties": {
          "insight_id": {
            "title": "Insight Id",
            "type": "string"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "insight_id",
          "metric_id",
          "title",
          "statement"
        ],
        "title": "DashboardInsight",
        "type": "object"
      },
      "DashboardKpi": {
        "additionalProperties": false,
        "properties": {
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Comparison"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence_interval_95": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceInterval"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "metric_id",
          "label",
          "value",
          "unit",
          "evidence_id"
        ],
        "title": "DashboardKpi",
        "type": "object"
      },
      "DashboardMerchant": {
        "additionalProperties": false,
        "properties": {
          "category_id": {
            "title": "Category Id",
            "type": "integer"
          },
          "category_title": {
            "title": "Category Title",
            "type": "string"
          },
          "merchant_key": {
            "title": "Merchant Key",
            "type": "string"
          },
          "terminal_count": {
            "title": "Terminal Count",
            "type": "integer"
          }
        },
        "required": [
          "merchant_key",
          "category_id",
          "category_title",
          "terminal_count"
        ],
        "title": "DashboardMerchant",
        "type": "object"
      },
      "DashboardMeta": {
        "additionalProperties": false,
        "properties": {
          "calculation_version": {
            "title": "Calculation Version",
            "type": "string"
          },
          "data_freshness": {
            "title": "Data Freshness",
            "type": "string"
          },
          "date_from": {
            "format": "date",
            "title": "Date From",
            "type": "string"
          },
          "date_to": {
            "format": "date",
            "title": "Date To",
            "type": "string"
          },
          "merchant_key": {
            "title": "Merchant Key",
            "type": "string"
          },
          "partial_data": {
            "title": "Partial Data",
            "type": "boolean"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "timezone": {
            "const": "Asia/Tehran",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "merchant_key",
          "date_from",
          "date_to",
          "timezone",
          "source_kind",
          "data_freshness",
          "partial_data",
          "calculation_version"
        ],
        "title": "DashboardMeta",
        "type": "object"
      },
      "DashboardResponse": {
        "additionalProperties": false,
        "properties": {
          "adjusted_fee_notice": {
            "title": "Adjusted Fee Notice",
            "type": "string"
          },
          "amount_bands": {
            "items": {
              "$ref": "#/components/schemas/AmountBandRow"
            },
            "title": "Amount Bands",
            "type": "array"
          },
          "insights": {
            "items": {
              "$ref": "#/components/schemas/DashboardInsight"
            },
            "title": "Insights",
            "type": "array"
          },
          "kpis": {
            "items": {
              "$ref": "#/components/schemas/DashboardKpi"
            },
            "title": "Kpis",
            "type": "array"
          },
          "latency": {
            "$ref": "#/components/schemas/LatencySection"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/LifecycleSection"
          },
          "merchant": {
            "$ref": "#/components/schemas/DashboardMerchant"
          },
          "meta": {
            "$ref": "#/components/schemas/DashboardMeta"
          },
          "metric_registry": {
            "items": {
              "$ref": "#/components/schemas/MetricRegistryItem"
            },
            "title": "Metric Registry",
            "type": "array"
          },
          "opportunity": {
            "$ref": "#/components/schemas/OpportunitySection"
          },
          "paid_not_verified": {
            "$ref": "#/components/schemas/PaidNotVerifiedSection"
          },
          "peer": {
            "$ref": "#/components/schemas/PeerSection"
          },
          "psp_code_coverage": {
            "$ref": "#/components/schemas/PspCodeCoverage"
          },
          "psp_codes": {
            "items": {
              "$ref": "#/components/schemas/PspCodeRow"
            },
            "title": "Psp Codes",
            "type": "array"
          },
          "repeat": {
            "$ref": "#/components/schemas/RepeatSection"
          },
          "retry": {
            "$ref": "#/components/schemas/RetrySection"
          },
          "revenue_decomposition": {
            "$ref": "#/components/schemas/RevenueDecomposition"
          },
          "trend": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "title": "Trend",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "merchant",
          "kpis",
          "lifecycle",
          "trend",
          "amount_bands",
          "revenue_decomposition",
          "retry",
          "repeat",
          "paid_not_verified",
          "latency",
          "psp_codes",
          "psp_code_coverage",
          "peer",
          "opportunity",
          "insights",
          "metric_registry",
          "adjusted_fee_notice"
        ],
        "title": "DashboardResponse",
        "type": "object"
      },
...
      "EvidenceRecordsResponse": {
        "additionalProperties": false,
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "grain": {
            "enum": [
              "session",
              "attempt",
              "customer_within_merchant"
            ],
            "title": "Grain",
            "type": "string"
          },
          "insight_id": {
            "title": "Insight Id",
            "type": "string"
          },
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SessionEvidenceRecord"
                },
                {
                  "$ref": "#/components/schemas/AttemptEvidenceRecord"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "merchant_key": {
            "title": "Merchant Key",
            "type": "string"
          },
          "page": {
            "minimum": 1,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100,
            "minimum": 1,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "insight_id",
          "merchant_key",
          "page",
          "page_size",
          "total",
          "items",
          "columns",
          "grain",
          "limitations"
        ],
        "title": "EvidenceRecordsResponse",
        "type": "object"
      },
...
      "EvidenceResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "$ref": "#/components/schemas/InsightEvidence"
          }
        },
        "required": [
          "evidence"
        ],
        "title": "EvidenceResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LatencySection": {
        "additionalProperties": false,
        "properties": {
          "definition": {
            "title": "Definition",
            "type": "string"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "init_median_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Init Median Ms"
          },
          "init_p90_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Init P90 Ms"
          },
          "init_p95_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Init P95 Ms"
          },
          "init_sample_size": {
            "title": "Init Sample Size",
            "type": "integer"
          },
          "psp_sample_size": {
            "title": "Psp Sample Size",
            "type": "integer"
          },
          "switch_code_sample_size": {
            "title": "Switch Code Sample Size",
            "type": "integer"
          },
          "verify_evidence_id": {
            "title": "Verify Evidence Id",
            "type": "string"
          },
          "verify_median_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify Median Ms"
          },
          "verify_p90_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify P90 Ms"
          },
          "verify_p95_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify P95 Ms"
          },
          "verify_sample_size": {
            "title": "Verify Sample Size",
            "type": "integer"
          }
        },
        "required": [
          "psp_sample_size",
          "switch_code_sample_size",
          "init_sample_size",
          "init_median_ms",
          "init_p90_ms",
          "init_p95_ms",
          "verify_sample_size",
          "verify_median_ms",
          "verify_p90_ms",
          "verify_p95_ms",
          "definition",
          "evidence_id",
          "verify_evidence_id"
        ],
        "title": "LatencySection",
        "type": "object"
      },
      "LifecycleSection": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "grain": {
            "const": "session",
            "title": "Grain",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/LifecycleStage"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "stages",
          "grain",
          "evidence_id"
        ],
        "title": "LifecycleSection",
        "type": "object"
      },
...
      "MetricRegistryItem": {
        "additionalProperties": false,
        "properties": {
          "business_question": {
            "title": "Business Question",
            "type": "string"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "denominator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Denominator"
          },
          "description_fa": {
            "title": "Description Fa",
            "type": "string"
          },
          "exclusions": {
            "items": {
              "type": "string"
            },
            "title": "Exclusions",
            "type": "array"
          },
          "filters": {
            "items": {
              "type": "string"
            },
            "title": "Filters",
            "type": "array"
          },
          "formula": {
            "title": "Formula",
            "type": "string"
          },
          "grain": {
            "enum": [
              "attempt",
              "session",
              "customer_within_merchant",
              "terminal",
              "merchant",
              "category",
              "time_period"
            ],
            "title": "Grain",
            "type": "string"
          },
          "metric_id": {
            "title": "Metric Id",
            "type": "string"
          },
          "minimum_sample_rule": {
            "title": "Minimum Sample Rule",
            "type": "string"
          },
          "name_fa": {
            "title": "Name Fa",
            "type": "string"
          },
          "null_policy": {
            "title": "Null Policy",
            "type": "string"
          },
          "numerator": {
            "title": "Numerator",
            "type": "string"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "rag_thresholds": {
            "$ref": "#/components/schemas/RagThresholds"
          },
          "source_columns": {
            "items": {
              "type": "string"
            },
            "title": "Source Columns",
            "type": "array"
          },
          "tests": {
            "items": {
              "type": "string"
            },
            "title": "Tests",
            "type": "array"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "weighting": {
            "title": "Weighting",
            "type": "string"
          }
        },
        "required": [
          "metric_id",
          "name_fa",
          "description_fa",
          "business_question",
          "decision",
          "grain",
          "formula",
          "numerator",
          "denominator",
          "source_columns",
          "filters",
          "exclusions",
          "null_policy",
          "unit",
          "currency",
          "weighting",
          "minimum_sample_rule",
          "owner",
          "version",
          "rag_thresholds",
          "tests"
        ],
        "title": "MetricRegistryItem",
        "type": "object"
      },
      "OpportunitySection": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "formula": {
            "title": "Formula",
            "type": "string"
          },
          "is_forecast": {
            "const": false,
            "title": "Is Forecast",
            "type": "boolean"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "value",
          "unit",
          "is_forecast",
          "formula",
          "evidence_id"
        ],
        "title": "OpportunitySection",
        "type": "object"
      },
      "PaidNotVerifiedSection": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          },
          "sessions": {
            "title": "Sessions",
            "type": "integer"
          }
        },
        "required": [
          "sessions",
          "rate",
          "evidence_id"
        ],
        "title": "PaidNotVerifiedSection",
        "type": "object"
      },
...
      "PspCodeCoverage": {
        "additionalProperties": false,
        "properties": {
          "coded_attempts": {
            "title": "Coded Attempts",
            "type": "integer"
          },
          "coverage_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Rate"
          },
          "eligible_psp_attempts": {
            "title": "Eligible Psp Attempts",
            "type": "integer"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "missing_code_attempts": {
            "title": "Missing Code Attempts",
            "type": "integer"
          }
        },
        "required": [
          "eligible_psp_attempts",
          "coded_attempts",
          "missing_code_attempts",
          "coverage_rate",
          "evidence_id"
        ],
        "title": "PspCodeCoverage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Deterministic, traceable merchant analytics at canonical payment grains.",
    "title": "نبض زرین — ZarinPal Merchant Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/dashboard": {
      "get": {
        "operationId": "dashboard_api_v1_dashboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "merchant_key",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_]+$",
              "title": "Merchant Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/insights/{insight_id}/evidence": {
      "get": {
        "operationId": "evidence_api_v1_insights__insight_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "insight_id",
            "required": true,
            "schema": {
              "title": "Insight Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence",
        "tags": [
          "evidence"
        ]
      }
    },
    "/api/v1/insights/{insight_id}/records": {
      "get": {
        "operationId": "records_api_v1_insights__insight_id__records_get",
        "parameters": [
          {
            "in": "path",
            "name": "insight_id",
            "required": true,
            "schema": {
              "title": "Insight Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRecordsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Records",
        "tags": [
          "evidence"
        ]
      }
    },
...
  },
...
}